# docs/openApi.json
{
 "components": {
  "schemas": {
   "Error": {
    "description": "Standard JSON error",
    "properties": {
     "message": {
      "type": "string"
     },
     "statusCode": {
      "maximum": 600,
      "minimum": 100,
      "type": "integer"
     }
    },
    "required": [
     "message",
     "statusCode"
    ],
    "type": "object"
   },
   "viseh-api_types_Chip": {
    "properties": {
     "data": {
      "$ref": "#/components/schemas/viseh-api_types_NfcData"
     },
     "keys": {
      "items": {
       "type": "string"
      },
      "nullable": true,
      "type": "array"
     },
     "version": {
      "type": "integer"
     }
    },
    "required": [
     "version",
     "keys",
     "data"
    ],
    "type": "object"
   },
   "viseh-api_types_ChipEncryptedV2": {
    "properties": {
     "data": {
      "type": "string"
     },
     "keys": {
      "items": {
       "type": "string"
      },
      "nullable": true,
      "type": "array"
     },
     "version": {
      "type": "integer"
     }
    },
    "required": [
     "version",
     "keys",
     "data"
    ],
    "type": "object"
   },
   "viseh-api_types_Conditions": {
    "properties": {
     "description": {
      "type": "string"
     },
     "name": {
      "type": "string"
     },
     "severity": {
      "type": "integer"
     }
    },
    "required": [
     "name",
     "severity",
     "description"
    ],
    "type": "object"
   },
...
   "viseh-api_types_MedUser": {
    "description": "User object",
    "nullable": true,
    "properties": {
     "email": {
      "description": "email address",
      "example": "gloriastehr@johnston.com",
      "type": "string"
     },
     "id": {
      "description": "unique user id",
      "example": 1,
      "type": "integer"
     },
     "name": {
      "description": "first and last name",
      "example": "Mack Jast",
      "type": "string"
     },
     "organisation": {
      "description": "medical organisation",
      "example": "OpenPlans",
      "type": "string"
     },
     "phone_number": {
      "description": "phone number",
      "example": "1-088-465-0498",
      "type": "string"
     },
     "role": {
      "description": "role",
      "enum": [
       "ambulance",
       "doctor"
      ],
      "example": "ambulance",
      "type": "string"
     }
    },
    "required": [
     ""
    ],
    "type": "object"
   },
   "viseh-api_types_NfcData": {
    "properties": {
     "ICE": {
      "type": "string"
     },
     "allergies": {
      "items": {
       "type": "string"
      },
      "nullable": true,
      "type": "array"
     },
     "blood-type": {
      "type": "string"
     },
     "conditions": {
      "items": {
       "$ref": "#/components/schemas/viseh-api_types_Conditions"
      },
      "nullable": true,
      "type": "array"
     },
     "date-of-birth": {
      "type": "string"
     },
     "dnr": {
      "type": "boolean"
     },
     "implants": {
      "items": {
       "type": "string"
      },
      "nullable": true,
      "type": "array"
     },
     "med-id": {
      "type": "string"
     },
     "medications": {
      "items": {
       "type": "string"
      },
      "nullable": true,
      "type": "array"
     },
     "name": {
      "type": "string"
     }
    },
    "required": [
     "med-id",
     "dnr",
     "blood-type",
     "name",
     "date-of-birth",
     "ICE",
     "allergies",
     "medications",
     "implants",
     "conditions"
    ],
    "type": "object"
   },
...
   "viseh-api_types_UserLogin": {
    "properties": {
     "email": {
      "type": "string"
     },
     "password": {
      "type": "string"
     }
    },
    "type": "object"
   }
  }
 },
 "info": {
  "title": "messages",
  "version": "0.0.0"
 },
 "openapi": "3.0.0",
 "paths": {
  "/dataProtection": {
   "get": {
    "description": "Data protection policy.",
    "responses": {
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   }
  },
...
  "/med/login": {
   "post": {
    "description": "Login.",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/viseh-api_types_UserLogin"
       }
      }
     }
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   }
  },
  "/med/mock/{count}": {
   "get": {
    "description": "Generate mock medical users.",
    "parameters": [
     {
      "description": "amount of medical users to mock",
      "in": "path",
      "name": "count",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   }
  },
  "/med/nfc/{med_id}": {
   "get": {
    "description": "Get nfc data by medical id.",
    "parameters": [
     {
      "description": "medical id of the user",
      "in": "path",
      "name": "med_id",
      "required": true,
      "schema": {
       "pattern": "\\d+",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/viseh-api_types_Chip"
        }
       }
      },
      "description": ""
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   }
  },
  "/med/register": {
   "post": {
    "description": "Register.",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/viseh-api_types_MedUser"
       }
      }
     }
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   }
  },
  "/med/users": {
   "get": {
    "description": "Get list of medical users.",
    "parameters": [
     {
      "description": "amount of items",
      "in": "query",
      "name": "items",
      "schema": {
       "type": "integer"
      }
     },
     {
      "description": "current page index",
      "in": "query",
      "name": "page",
      "schema": {
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/viseh-api_types_MedUser"
         },
         "nullable": true,
         "type": "array"
        }
       }
      },
      "description": ""
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   }
  },
  "/med/users/search": {
   "get": {
    "description": "Search for medical user.",
    "parameters": [
     {
      "description": "user email",
      "in": "query",
      "name": "email",
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/viseh-api_types_MedUser"
         },
         "nullable": true,
         "type": "array"
        }
       }
      },
      "description": ""
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   }
  },
  "/med/users/{id}": {
   "get": {
    "description": "Get medical user by uid.",
    "parameters": [
     {
      "description": "id of the user",
      "in": "path",
      "name": "uid",
      "required": true,
      "schema": {
       "pattern": "\\d+",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/viseh-api_types_MedUser"
        }
       }
      },
      "description": ""
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   }
  },
  "/med/users/{uid}": {
   "patch": {
    "description": "Patch medical user.",
    "parameters": [
     {
      "description": "id of the user",
      "in": "path",
      "name": "uid",
      "required": true,
      "schema": {
       "pattern": "\\d+",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/viseh-api_types_MedUser"
       }
      }
     }
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   },
   "put": {
    "description": "Update medical user.",
    "parameters": [
     {
      "description": "id of the user",
      "in": "path",
      "name": "uid",
      "required": true,
      "schema": {
       "pattern": "\\d+",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/viseh-api_types_MedUser"
       }
      }
     }
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   }
  },
  "/mock/{count}": {
   "get": {
    "description": "Generate mock users.",
    "parameters": [
     {
      "description": "amount of users to mock",
      "in": "path",
      "name": "count",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   }
  },
  "/nfc/{uid}": {
   "get": {
    "description": "Get nfc data by uid.",
    "parameters": [
     {
      "description": "id of the user",
      "in": "path",
      "name": "uid",
      "required": true,
      "schema": {
       "pattern": "\\d+",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/viseh-api_types_ChipEncryptedV2"
        }
       }
      },
      "description": ""
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   },
   "post": {
    "description": "Get user by uid.",
    "parameters": [
     {
      "description": "id of the user",
      "in": "path",
      "name": "uid",
      "required": true,
      "schema": {
       "pattern": "\\d+",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/viseh-api_types_Chip"
       }
      }
     }
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   }
  },
  "/privacy": {
   "get": {
    "description": "Privacy policy.",
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   }
  },
...
 }
}